clean up

+++ Numerics/Lie.py
@@ -5,21 +5,21 @@ def generator_x():
     return np.array([[0,0,0,1],
                      [0,0,0,0],
                      [0,0,0,0],
-                     [0,0,0,0]],dtype=np.float32)
+                     [0,0,0,0]],dtype=np.float64)
 
 
 def generator_y():
     return np.array([[0,0,0,0],
                      [0,0,0,1],
                      [0,0,0,0],
-                     [0,0,0,0]],dtype=np.float32)
+                     [0,0,0,0]],dtype=np.float64)
 
 
 def generator_z():
     return np.array([[0,0,0,0],
                      [0,0,0,0],
                      [0,0,0,1],
-                     [0,0,0,0]],dtype=np.float32)
+                     [0,0,0,0]],dtype=np.float64)
 
 
 # Rotation Around X
@@ -27,7 +27,7 @@ def generator_roll():
     return np.array([[0,0,0,0],
                      [0,0,-1,0],
                      [0,1,0,0],
-                     [0,0,0,0]],dtype=np.float32)
+                     [0,0,0,0]],dtype=np.float64)
 
 
 # Rotation Around Y
@@ -35,7 +35,7 @@ def generator_pitch():
     return np.array([[0,0,1,0],
                      [0,0,0,0],
                      [-1,0,0,0],
-                     [0,0,0,0]],dtype=np.float32)
+                     [0,0,0,0]],dtype=np.float64)
 
 
 # Rotation Around Z
@@ -43,4 +43,4 @@ def generator_yaw():
     return np.array([[0,-1,0,0],
                      [1,0,0,0],
                      [0,0,0,0],
-                     [0,0,0,0]],dtype=np.float32)
+                     [0,0,0,0]],dtype=np.float64)
